# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: API for WATT A PLAY 3.0 campus treasure hunt
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/team/login:
    post:
      operationId: teamLogin
      tags: [auth]
      summary: Claim a team captain session with one access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamLoginInput"
      responses:
        "200":
          description: Signed in team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid code, captain name required, or another device owns the session
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Sign in as an administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Signed in administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    AuthUser:
      type: object
      properties:
        id: { type: string }
        role: { type: string, enum: [TEAM, ADMIN] }
        name: { type: ["string", "null"] }
        teamId: { type: ["string", "null"] }
      required: [id, role, name, teamId]
    AuthResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/AuthUser" }
      required: [user]
    TeamLoginInput:
      type: object
      properties:
        code: { type: string, minLength: 8, maxLength: 16 }
        deviceId: { type: string, minLength: 8, maxLength: 100 }
        leaderName: { type: string, maxLength: 100 }
      required: [code, deviceId]
    AdminLoginInput:
      type: object
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
      required: [username, password]
...