# lib/api-zod/src/generated/types/adminLoginInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for WATT A PLAY 3.0 campus treasure hunt
 * OpenAPI spec version: 1.0.0
 */

export interface AdminLoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}