# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for WATT A PLAY 3.0 campus treasure hunt
 * OpenAPI spec version: 1.0.0
 */
import type { AuthUserRole } from './authUserRole';

export interface AuthUser {
  id: string;
  role: AuthUserRole;
  /** @nullable */
  name: string | null;
  /** @nullable */
  teamId: string | null;
}